Identify grid view filter controls automatically from HTML form
There is no longer any need to manually repeat and bind all the filter
controls in the script file. All <input>s and <select>s in the form are
now "bound" automatically. The name attribute from each element is
mapped to a function ("getter") that extracts the value of that form
control, and this map is passed as the postData option to jqGrid. jqGrid
then constructs query params by assigning the getter's return value to
the respective name.

Additional post-processing of the parameters can be specified using the
PARAM_TRANSFORMERS constant. Any functions specified here are hooked in
before returning the param value. The transformer functions are passed
the value extracted from the form control as the first parameter and are
expected to return the final parameter value.